# swagger.json
{  
   "swagger":"2.0",
   "info":{  
      "version":"v1",
      "title":"Bike Rental Service"
   },
   "paths":{  
      "/api/bikes":{  
         "get":{  
            "tags":[  
               "Bike"
            ],
            "operationId":"ApiBikesGet",
            "consumes":[  

            ],
            "produces":[  

            ],
            "parameters":[  
               {  
                  "name":"sortCriteria",
                  "in":"query",
                  "required":false,
                  "type":"string"
               }
            ],
            "responses":{  
               "200":{  
                  "description":"Success"
               }
            }
         },
         "post":{  
            "tags":[  
               "Bike"
            ],
            "operationId":"ApiBikesPost",
            "consumes":[  
               "application/json-patch+json",
               "application/json",
               "text/json",
               "application/*+json"
            ],
            "produces":[  

            ],
            "parameters":[  
               {  
                  "name":"bike",
                  "in":"body",
                  "required":false,
                  "schema":{  
                     "$ref":"#/definitions/Bike"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"Success"
               }
            }
         }
      },
      "/api/bikes/{id}":{  
         "put":{  
            "tags":[  
               "Bike"
            ],
            "operationId":"ApiBikesByIdPut",
            "consumes":[  
               "application/json-patch+json",
               "application/json",
               "text/json",
               "application/*+json"
            ],
            "produces":[  

            ],
            "parameters":[  
               {  
                  "name":"id",
                  "in":"path",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               },
               {  
                  "name":"newBike",
                  "in":"body",
                  "required":false,
                  "schema":{  
                     "$ref":"#/definitions/Bike"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"Success"
               }
            }
         },
         "delete":{  
            "tags":[  
               "Bike"
            ],
            "operationId":"ApiBikesByIdDelete",
            "consumes":[  

            ],
            "produces":[  

            ],
            "parameters":[  
               {  
                  "name":"id",
                  "in":"path",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{  
               "200":{  
                  "description":"Success"
               }
            }
         }
      },
      "/api/customers":{  
         "get":{  
            "tags":[  
               "Customer"
            ],
            "operationId":"ApiCustomersGet",
            "consumes":[  

            ],
            "produces":[  

            ],
            "parameters":[  
               {  
                  "name":"query",
                  "in":"query",
                  "required":false,
                  "type":"string"
               }
            ],
            "responses":{  
               "200":{  
                  "description":"Success"
               }
            }
         },
         "post":{  
            "tags":[  
               "Customer"
            ],
            "operationId":"ApiCustomersPost",
            "consumes":[  
               "application/json-patch+json",
               "application/json",
               "text/json",
               "application/*+json"
            ],
            "produces":[  

            ],
            "parameters":[  
               {  
                  "name":"customer",
                  "in":"body",
                  "required":false,
                  "schema":{  
                     "$ref":"#/definitions/Customer"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"Success"
               }
            }
         }
      },
      "/api/customers/rentals/{customerId}":{  
         "get":{  
            "tags":[  
               "Customer"
            ],
            "operationId":"ApiCustomersRentalsByCustomerIdGet",
            "consumes":[  

            ],
            "produces":[  

            ],
            "parameters":[  
               {  
                  "name":"customerId",
                  "in":"path",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{  
               "200":{  
                  "description":"Success"
               }
            }
         }
      },
      "/api/customers/{id}":{  
         "put":{  
            "tags":[  
               "Customer"
            ],
            "operationId":"ApiCustomersByIdPut",
            "consumes":[  
               "application/json-patch+json",
               "application/json",
               "text/json",
               "application/*+json"
            ],
            "produces":[  

            ],
            "parameters":[  
               {  
                  "name":"id",
                  "in":"path",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               },
               {  
                  "name":"newCustomer",
                  "in":"body",
                  "required":false,
                  "schema":{  
                     "$ref":"#/definitions/Customer"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"Success"
               }
            }
         },
         "delete":{  
            "tags":[  
               "Customer"
            ],
            "operationId":"ApiCustomersByIdDelete",
            "consumes":[  

            ],
            "produces":[  

            ],
            "parameters":[  
               {  
                  "name":"id",
                  "in":"path",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{  
               "200":{  
                  "description":"Success"
               }
            }
         }
      },
      "/api/rentals/start":{  
         "get":{  
            "tags":[  
               "Rental"
            ],
            "operationId":"ApiRentalsStartGet",
            "consumes":[  

            ],
            "produces":[  

            ],
            "parameters":[  
               {  
                  "name":"customerId",
                  "in":"query",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               },
               {  
                  "name":"bikeId",
                  "in":"query",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{  
               "200":{  
                  "description":"Success"
               }
            }
         }
      },
      "/api/rentals/end/{rentalId}":{  
         "get":{  
            "tags":[  
               "Rental"
            ],
            "operationId":"ApiRentalsEndByRentalIdGet",
            "consumes":[  

            ],
            "produces":[  

            ],
            "parameters":[  
               {  
                  "name":"rentalId",
                  "in":"path",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{  
               "200":{  
                  "description":"Success"
               }
            }
         }
      },
      "/api/rentals/pay/{rentalId}":{  
         "get":{  
            "tags":[  
               "Rental"
            ],
            "operationId":"ApiRentalsPayByRentalIdGet",
            "consumes":[  

            ],
            "produces":[  

            ],
            "parameters":[  
               {  
                  "name":"rentalId",
                  "in":"path",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{  
               "200":{  
                  "description":"Success"
               }
            }
         }
      },
      "/api/rentals/unpaid":{  
         "get":{  
            "tags":[  
               "Rental"
            ],
            "operationId":"ApiRentalsUnpaidGet",
            "consumes":[  

            ],
            "produces":[  

            ],
            "responses":{  
               "200":{  
                  "description":"Success"
               }
            }
         }
      },
      "/api/rentals":{  
         "get":{  
            "tags":[  
               "Rental"
            ],
            "operationId":"ApiRentalsGet",
            "consumes":[  

            ],
            "produces":[  

            ],
            "responses":{  
               "200":{  
                  "description":"Success"
               }
            }
         }
      }
   },
   "definitions":{  
      "Bike":{  
         "required":[  
            "brand",
            "purchaseDate",
            "firstHourPrice",
            "additionalHourPrice"
         ],
         "type":"object",
         "properties":{  
            "id":{  
               "format":"int32",
               "type":"integer"
            },
            "brand":{  
               "maxLength":25,
               "type":"string"
            },
            "purchaseDate":{  
               "format":"date-time",
               "type":"string"
            },
            "notes":{  
               "maxLength":1000,
               "type":"string"
            },
            "lastServiceDate":{  
               "format":"date-time",
               "type":"string"
            },
            "firstHourPrice":{  
               "format":"double",
               "minimum":0,
               "pattern":"\\d+(\\.\\d{1,2})?",
               "type":"number"
            },
            "additionalHourPrice":{  
               "format":"double",
               "minimum":1,
               "pattern":"\\d+(\\.\\d{1,2})?",
               "type":"number"
            },
            "category":{  
               "format":"int32",
               "enum":[  
                  1,
                  2,
                  3,
                  4
               ],
               "type":"integer"
            }
         }
      },
      "Customer":{  
         "required":[  
            "firstName",
            "lastName",
            "birthdate",
            "street",
            "zipCode",
            "town"
         ],
         "type":"object",
         "properties":{  
            "id":{  
               "format":"int32",
               "type":"integer"
            },
            "gender":{  
               "format":"int32",
               "enum":[  
                  1,
                  2,
                  3
               ],
               "type":"integer"
            },
            "firstName":{  
               "maxLength":50,
               "type":"string"
            },
            "lastName":{  
               "maxLength":75,
               "type":"string"
            },
            "birthdate":{  
               "format":"date-time",
               "type":"string"
            },
            "street":{  
               "maxLength":50,
               "type":"string"
            },
            "houseNumber":{  
               "format":"int32",
               "maximum":2147483647,
               "minimum":0,
               "type":"integer"
            },
            "zipCode":{  
               "format":"int32",
               "maximum":2147483647,
               "minimum":0,
               "type":"integer"
            },
            "town":{  
               "maxLength":50,
               "type":"string"
            }
         }
      }
   },
   "securityDefinitions":{  

   }
}